refactor: move provider to separate folder
